refactor: modularize switch setup helpers into builder modules
- Created `custom_components/meraki_ha/switch/builders/` subpackage.
- Extracted entity setup logic from `setup_helpers.py` into specific modules:
  - `firewall.py`: L3 firewall rule switches.
  - `traffic_shaping.py`: Traffic shaping switches.
  - `vpn.py`: Site-to-site VPN switches.
  - `vlan.py`: VLAN DHCP switches.
  - `ssid.py`: SSID availability and broadcast switches.
  - `camera.py`: Camera analytics switches.
  - `mt40.py`: MT40 power outlet switches.
- Refactored `setup_helpers.py` to orchestrate setup calls to these builders.
- Significantly reduced ACL complexity of `setup_helpers.py`.
- Verified changes with `tests/switch/` suite.
